Parse args in utils_shared
